Remap j and k to move into wrapped lines
Normally j and k move over wrapped lines and onto the next actual line
in the file. If a line is wrapped because the window is too narrow, the
wrapped portion of the lines will be skipped and you need to press gj/gk
to move among the wrapped portions. Normally we want to move into
wrapped lines, so remap j and k to do this.